+++ bongo/fmt/detail/printer.cpp
@@ -8,15 +8,15 @@
 
 namespace bongo::fmt::detail {
 
-auto printer::width() const -> std::pair<int, bool> {
+auto printer::width() const -> std::pair<long, bool> {
   return {fmt_.wid, fmt_.flags.wid_present};
 }
 
-auto printer::precision() const -> std::pair<int, bool> {
+auto printer::precision() const -> std::pair<long, bool> {
   return {fmt_.prec, fmt_.flags.prec_present};
 }
 
-auto printer::flag(int b) const -> bool {
+auto printer::flag(long b) const -> bool {
   switch (b) {
   case '-':
     return fmt_.flags.minus;
@@ -33,7 +33,7 @@ auto printer::flag(int b) const -> bool {
   }
 }
 
-auto printer::write(std::span<uint8_t const> b) -> std::pair<int, std::error_code> {
+auto printer::write(std::span<uint8_t const> b) -> std::pair<long, std::error_code> {
   buf_.write(b);
   return {b.size(), nil};
 }
